Refactoring
Removed express-validation lib
Update Readme.md
removed some commented code
