+++ package.json
@@ -32,7 +32,6 @@
     "body-parser": "^1.19.1",
     "dotenv": "^10.0.0",
     "express": "^4.17.2",
-    "express-validator": "^6.14.0",
     "jsonwebtoken": "^8.5.1",
     "mysql2": "^2.3.3",
     "sequelize": "^7.0.0-alpha.2",
